A Poly object can be used as a function
Just some syntactic sugar. For julia version 0.4 or newer,
the `call` method can be overloaded and can be convenient when
using Poly objects.

This pull request, adds the new `call` method if Julia version is
0.4 or newer.

Added a test in test/runtests.jl that only gets executed if
version is 0.4 or newer.

The tests passed in both version 0.3 and 0.4.