Change indexing of `Poly`s
Changed indexing (both `getindex` and `setindex!`) of `Poly`s to add
support for `::Colon` and to have compatibility with `base/array.jl`.
The compatibility is about returning the changed `p::Poly` after
`setindex!`s. Moreover, `setindex!` had an array comprehension before,
resulting in an efficiency penalty (creating a temporary, and not using
it at all).