Change formulation "overloading"
I have changed a formulation under "Arithmetic" that said that the basic arithmetic operators are overloaded. If I am not mistaken, this is not what it is called in Julia, but it is what it is called in many other languages. I think that what was meant is that methods are added to the basic arithmetic operators, and so, I have changed the formulation accordingly.